+++ src/main/java/command/Command.java
@@ -1,3 +1,7 @@
+package command;
+import duke.Ui;
+import duke.TaskList;
+
 public abstract class Command {
 
     private TaskList taskList;
