+++ src/main/java/command/eventCommand.java
@@ -1,3 +1,8 @@
+package command;
+import duke.Ui;
+import duke.TaskList;
+
+import task.Event;
 import exception.EmptyInputException;
 import exception.EmptyTimeException;
 import exception.InvalidDateTimeException;
