+++ src/main/java/command/markCommand.java
@@ -1,18 +1,22 @@
-public class deleteCommand extends Command {
+package command;
+import duke.Ui;
+import duke.TaskList;
+
+public class markCommand extends Command {
 
     private TaskList taskList;
     private Ui ui;
 
-    public deleteCommand(TaskList taskList, Ui ui) {
+    public markCommand(TaskList taskList, Ui ui) {
         super(taskList, ui);
     }
-
     @Override
+
     public void execute(TaskList taskList, Ui ui) throws Exception {
         String input = ui.getInput();
         String indexStr = input.split(" ")[1];
         int position = Integer.parseInt(indexStr) -1;
-        taskList.delete(position);
+        taskList.mark(position);
     }
 
     @Override
