+++ src/main/java/command/unmarkCommand.java
@@ -1,21 +1,23 @@
-public class markCommand extends Command {
+package command;
+import duke.Ui;
+import duke.TaskList;
+
+public class unmarkCommand extends Command {
 
     private TaskList taskList;
     private Ui ui;
 
-    public markCommand(TaskList taskList, Ui ui) {
+    public unmarkCommand(TaskList taskList, Ui ui) {
         super(taskList, ui);
     }
-    @Override
 
     public void execute(TaskList taskList, Ui ui) throws Exception {
         String input = ui.getInput();
         String indexStr = input.split(" ")[1];
         int position = Integer.parseInt(indexStr) -1;
-        taskList.mark(position);
+        taskList.unmark(position);
     }
 
-    @Override
     public boolean isExit() {
         return false;
     }
