+++ src/main/java/duke/Actions.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum Actions {
     BYE,
     LIST,
