+++ src/main/java/duke/Parser.java
@@ -1,4 +1,5 @@
-import exception.*;
+package duke;
+import command.*;
 
 public class Parser {
     private static TaskList taskList;
