+++ src/main/java/task/Event.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
